requirements.txt
openai>=1.0.0
streamlit>=1.28.0
python-dotenv>=1.0.0
pulp>=2.7.0
pandas>=2.0.0
numpy>=1.24.0
python-dateutil>=2.8.2
networkx>=3.0
matplotlib>=3.7.0
